Update tools.go
Added the rake function. it needs a final return argument in the NLP function, other than that the v.1 for producing new ghostwritten text based on the wikipaedia dump files just needs to utilize the given functions (arguments being article titles) and apply an A* algorithm or similar to traverse the articles' links. - Hmm, from here I planned to implement the naive bayes, tf-idf, and pagerank algorithms, where pagerank would definitely add an improvement. v.2 should communicate with a P2P network to share information (to save processing power) and make an interactive chat/learn bot or matrix-like terminal with keyword-based coloring. - There are also the translation option, rivescript, aiml for the keywords and sentence construction. And don't forget the perceptron based tagger if unknown writing styles or languages are included.