Update tools.go
Removed the pagerank function (apparently one has to be granted allowance by Google to use the pagerank algorithm) and replaced it with linear regression. Postprocess the Wikipedia articles ReadIndex_StageOne and .+_StageTwo first read the index of leaving articles from a base, that is given by user input, and then in 2nd stage reads the actual data from the articles (in compressed and encoded form).
TODO: write the main function, make sure everything compiles, benchmark it, extend it.