Update tools.go
This compiles, and the methods and types are ready for use (no given example though, you are welcome to write to me though), the only functionality is applying linear regression on nouns in articles and storing articles in a gzip compressed matter (and gob encoded metadata - zscore and extremum of nouns). Do mind the Wikipedia articles have to be stored separately, and I do not even remember if the code specifies where to look for the Wikipedia (bzip compressed files). The program takes you have the 26 Wikipedia files in compressed format. One thing; none of the methods specify what file a given article is in, this is going to have to be specified by oneself in the main function.