Update tools.go
If in doubt see comments (search for "TODO", "//"), A large part of the design structure is finished, and the linear regression is largely finished too (for the Wikipedia articles). What XML parser that needs to be implemented is the StackExchange XML's, and a design/data structure for these xml tags. Before StackExchange is implemented a the WriteTxt can be utilized with the Wikipedia pages (and an AIML chatbot with Bloom's keywords for taxonomical learning of n-depth of words and/or links (hence sentences are used here too for n-depth based learning)).