Create tools-new(2).go
Slightly modified tools.go: Decompressing of the wikipedia bzip files excluded (because indexing is apparently not well supported, all the content leading up to an article has to be decompressed instead of using a Seek method to a file offset), the indexing of wikipedia articles written in a separate source file. (index_wiki.go, uploaded after this)
This source file is still not complete, it - needs a concurrency mechanism by the sync package (wg.Add and Done methods) in the Read function, and then a method call from RankWords and RankArticles in the end of the main function.