Create index_wiki.go
I believe it should work, it did, it does not exceed 1.5GB of RAM (OS etc. memory included), but after using it I added a function call and have not tested since. - If you intend to use it in a Linux system rather merger the files together alphabetically because Linux systems do not, by default, support opening more than 1000 files. (~1050 files will be created for the index, non-latin characters excluded). The max open file limit can in Linux be changed and the ranker program modified to limit amount of open files though.