Create sentenizer.go
Right now it uses 2 external github repositories, optimally it should not, but they work. - but for the sake of code clarity a Penn Treebank parser could be used to extract sentences, or a comma/punct parser. - but the trained sentencer does seem more attractive on a range of unknown texts.
Also, I suggest not to use this code as is on a PC, rather it should be modified in the goroutine at line 190 as one likes.
I tried, and suddenly a half a dozen MB turned into 5+GB.

Obs: Remember to modify the paths to os.Create and/or os.Open etc. to your own liking, maybe by using the flag package in the golang stdlib.