Create tool.go
This file is to read an index of wikipedia (enwiki-latest-articles.xml from https://dumps.wikimedia.org/enwiki/latest/ ), to read single articles sorted by relevance to a root node. The code uses a CSV file with the articles' name and Pagerank from https://data.world/falcon/wikipedia-pagerank-20160804 ), the code is also linked from the website. The software made for the CSV file uses Google Colab, because a PC has not sufficient memory to process Wikipedia with Pagerank with a depth of 7 or epsilon of 0.15.