Create sample-rake-text.go
It is to use the RAKE algorithm to rank sentences by their relevance to the article. It seems fairly simple and justifiable to use in pair with a pagerank dataset of wikipedia, otherwise the TF-IDF might have been better, and still TF-IDF could be used for analysing the datasets further at the expense of computing time. The sentences will be paired with nouns from a text generated by