Create Noun-viterbi-processor.go
This is to read in a file with sentences in it, whether or not it has been written with special keywords like like xml. It pairs nouns in n-grams, 1, 2, 3, and sorts them in descending order to retrieve most used pairs of nouns faster. Linear regression is applied to the ngrams to weigh, to give them a value, according to their frequency (a hash table seems smart to use here to store a set of regression values between 0-1).